refactor(34-T7): 測試結構歸位 — 6 孤兒測試移入正確目錄 + conftest 分層
- test_actress_profile.py, test_new_scrapers.py, test_api_config.py → integration/
- test_scraper_partial.py → unit/
- test_scrapers.py → smoke/
- test_smoke.py 刪除（已被 smoke/test_scraper_live.py 覆蓋）
- 新建 tests/smoke/conftest.py（smoke server 管理邏輯從根 conftest 移入）
- unit/conftest.py 新增 client + make_mock_search_jav fixtures
- 移除 test_scanner_sqlite.py 重複 temp_db、根 conftest 重複 client
- 928 passed, 0 failed
