+++ tests/integration/test_new_scrapers.py
