+++ tests/unit/test_scanner_sqlite.py
@@ -5,16 +5,7 @@ import os
 from pathlib import Path
 
 from core.gallery_scanner import VideoScanner
-from core.database import init_db, VideoRepository
-
-
-@pytest.fixture
-def temp_db():
-    """建立臨時資料庫"""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        db_path = Path(tmpdir) / "test.db"
-        init_db(db_path)
-        yield db_path
+from core.database import VideoRepository
 
 
 @pytest.fixture
